+++ HDiSED/src/main/java/pl/woleszko/polsl/maths/impl/TankDataExtractor.java
@@ -25,10 +25,10 @@ public class TankDataExtractor extends DataExtractor {
 
 	public HashMap<Long, Double> getVolumeTotals(Period period) {
 		HashMap<Long, Double> totals = new HashMap<Long, Double>();
-		HashMap<Integer, Times> times = splitDates(period);
-		Set<Integer> keys = times.keySet();
+		HashMap<Long, Times> times = splitDates(period);
+		Set<Long> keys = times.keySet();
 		
-		for (Integer key : keys) {
+		for (Long key : keys) {
 			for (TankMeasuresEntity entity : entities) {
 				if (!totals.containsKey(entity.getTankId()))
 					totals.put(entity.getTankId(), (double) 0);
